add a simple unit test for the Dictionary
This is used to reproduce a bug in the Dictionary where consecutive
streams (that have a non-fully-high keep) are not processed correctly.
In particular, the second ids stream does not produce any value out.
This is due to a mismatch of ids in the internal tags used by the
dictionary, caused by a missing flush on some components.